Add hyperslab tests
